+++ packages/analytics-client-common/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.52](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-client-common@2.4.51...@amplitude/analytics-client-common@2.4.52) (2026-06-30)
+
+**Note:** Version bump only for package @amplitude/analytics-client-common
+
+
+
+
+
+## [2.4.51](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-client-common@2.4.50...@amplitude/analytics-client-common@2.4.51) (2026-06-25)
+
+**Note:** Version bump only for package @amplitude/analytics-client-common
+
+
+
+
+
 ## [2.4.50](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-client-common@2.4.49...@amplitude/analytics-client-common@2.4.50) (2026-06-11)
 
 **Note:** Version bump only for package @amplitude/analytics-client-common